+++ readme/RELEASE_NOTES.md
@@ -1,9 +1,7 @@
 _Note:_ When running your tests with this version for the first time, you will see that your snapshot files are being 
 changed because we introduced a new header field. You can and safely check in these changes into your SCM.
 
-* Improve orphan detection to report less false positives
-* Multiple JavaDoc improvements
-* Improve and simplify internal interface and test framework integration
+* Remove Target `ElementType.METHOD` from `@EnableSnapshotTests`
 
 Maven Central coordinates for this release: